Add files via upload
Number guessing game
The number guessing game is based on a concept where player guesses a number between a range(1 to
1000). We have to keep guessing the number until we find the computers number. If player guesses the
exact number then player wins else player looses the game. Each time we guess the wrong number, the
computer tell us whether the number is higher or lower than the random number generated by the
computer if we succeeded in guessing the number it will generate a green color. Since this game
provides limited attempts, so, player must guess the number with the limited attempts, else will lose the
game. As each guess is input, the background color will change to either red or blue. Red indicates the
user is getting “ warmer”, and blue indicates the user is getting “colder”. if the user gets the correct
answer it indicates the green color.
1.OBJECTIVE OF THE PROJECT
The main objective of this project is guessing a number is based on the concept of player guesses a
number and there are some rules to follow Number guessing game:
1) You must enter only valid integer within the specified range.
2) You will be provided unlimited attempts to guess the number.
3) You cannot leave the game, once started.
4) If the guess number is less than the Target value then it will display red color in background.
5) If the guess number is greater than the Target value it will display blue color in background.
6) If the guess number is equal to the Target value it will display green color in background.
7) If the entered number is less than or greater than the required number, then player gets the message
(hint) to proceed further either in up or down range.
2.USED TECHNOLOGY
We have used the technology java.java is a high level, class based, Object oriented Programming
language that is designed to have as few implementations dependencies as possible. In our project we
used like java Swings which includes the topics like JLabel, JFrame, Action Listener, Action Performed
etc. and Swing is a GUI Widget ToolKit for java.It is a part of oracle java foundation classes an API
provides a graphical user interface for java programming. Swing was developed to provide a more
sophisticated set of GUI components then the earlier Abstract Window Toolkit.
3.DEFINATIONS OF THE ITEMS
Class:
A class is a basic building block. It can be defined as template that describes the data and behaviour
associated with the class instantiation. Instantiating is a class is to create an object (variable) of that
class that can be used to access the member variables and methods of the class.
Object:
An object is a member of Java class. It is also known as instance of the class. objects are created at run
time. In simple words, object is a real word entity.
JLabel :
Is a class of java Swing . JLabel is used to display a short string or an image icon. JLabel can display
text, image or both . JLabel is only a display of text or image and it cannot get focus . JLabel is
inactive to input events such a mouse focus or keyboard focus.
JTextField :
Is a lightweight component that allows the editing of a single line of text. For information on and
examples of using text fields, see How to Use Text Fields in The Java Tutorial. JTextField is intended
to be source-compatible with java. awt. TextField where it is reasonable to do so.
Action Listener:
The listener interface for receiving action events. The class that is interested in processing an action
event implements this interface, and the object created with that class is registered with a component,
using the component's add ActionListener method.
Action Performed:
An action event occurs, whenever an action is performed by the user. Examples: When the user clicks a button, chooses a menu item, presses Enter in a text field. The result is that an actionPerformed
message is sent to all action listeners that are registered on the relevant component.
Event handling:
Event Handling is the mechanism that controls the event and decides what should happen if an event
occurs. This mechanism have the code which is known as event handler that is executed when an
event occurs. Java Uses the Delegation Event Model to handle the events.
Exception handling:
Exception Handling is a mechanism to handle runtime errors such as ClassNotFoundException,
IOException, SQLException, RemoteException, etc.
Awt:
AWT stands for Abstract Window Toolkit. It is a platform-dependent API to develop GUI (Graphical
User Interface) or window-based applications in Java. It was developed by Sun Microsystem In 1995.
It is heavy-weight in use because it is generated by the system's host operating system.
Swing:
Java Swing is a lightweight Java graphical user interface (GUI) widget toolkit that includes a rich set
of widgets. It is part of the Java Foundation Classes (JFC) and includes several packages for
developing rich desktop applications in Java.